+++ src/isthmuslib/__init__.py
@@ -1,4 +1,7 @@
-__version__ = '0.0.24'
+__version__ = '0.0.28'
 
 from .isthmuslib_plots import *
 from .isthmuslib_helpers import *
+from .config import *
+from .data_series import *
+from .utils import *